+++ src/gpio.rs
@@ -70,6 +70,7 @@ mod dynamic;
 pub use dynamic::{Dynamic, DynamicPin};
 mod hal_02;
 mod hal_1;
+pub mod outport;
 
 pub use embedded_hal::digital::v2::PinState;